nsmq-companion-backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.5.0
anyascii==0.3.2
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
bcrypt==4.0.1
bidict==0.23.1
black==23.7.0
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.34.113
botocore==1.34.113
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cryptography==41.0.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
dependency-injector==4.41.0
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
einops==0.8.0
email-validator==2.0.0.post2
encodec==0.1.1
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-cli==0.0.4
filelock==3.14.0
flake8==6.1.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.5.0
g2pkk==0.1.2
grpcio==1.64.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
isort==5.12.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
jiwer==3.0.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.42.0
Mako==1.3.5
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.22.0
onnx==1.16.1
onnxruntime==1.18.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.9.5
packaging==23.1
pandas==1.5.3
passlib==1.7.4
pathspec==0.11.2
phonenumbers==8.13.23
pillow==10.3.0
platformdirs==3.10.0
pooch==1.8.1
preshed==3.0.9
protobuf==5.27.0
psutil==5.9.8
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydub==0.25.1
pyflakes==3.1.0
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.2
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.11.4
sentence-transformers==3.0.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
starlette==0.27.0
structlog==23.2.0
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.4
trainer==0.0.36
transformers==4.41.1
TTS==0.22.0
typeguard==4.3.0
typer==0.9.4
typing_extensions==4.12.0
tzlocal==5.2
ujson==5.8.0
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
wasabi==1.1.2
watchfiles==0.20.0
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.0
